Do not lose virtual key code while handling KeyDown events with IME disabled
Related:
https://gitlab.com/freepascal.org/fpc/source/-/merge_requests/1138#note_2734665541
