+++ WinPort/src/Backend/WX/wxMain.cpp
@@ -1,6 +1,7 @@
 #include "wxMain.h"
 #include <dlfcn.h>
 #include "../NotifySh.h"
+#include "wxWinTranslations.h"
 
 #define AREAS_REDUCTION
 
@@ -1617,7 +1618,7 @@ void WinPortPanel::OnChar( wxKeyEvent& event )
 		ir.Event.KeyEvent.wRepeatCount = 1;
 		// we can not determine correct VirtualKeyCode value here because of
 		// https://github.com/wxWidgets/wxWidgets/issues/25379
-		ir.Event.KeyEvent.wVirtualKeyCode = VK_NONAME;
+		ir.Event.KeyEvent.wVirtualKeyCode = wxKeyCode2WinKeyCode(_key_tracker.LastKeydown().GetKeyCode());
 		if (event.GetUnicodeKey() <= 0x7f) { 
 			if (_key_tracker.LastKeydown().GetTimestamp() == event.GetTimestamp()) {
 				wx2INPUT_RECORD irx(TRUE, _key_tracker.LastKeydown(), _key_tracker);
